add solution to problem CF gym problem 100571B.
Solution based partially on partial sums logic, but mostly on decomposition of function calls if the problem was simulated.

To solve the problem in O(N), the logic is based on maintaining an array
with the quantity of f(3) calls for each position, and an array of
vectors, with the list of f(x) calls that end in that position.

Example:

query = [3, 7]

x[3] += f(1)
x[4] += f(2)
partial_insert[5]++;
partial_remove[7].push_back(5); // Because the call in this position
will be f(5)

This logic is based on the fact that you can apply the f(x) to an
auxiliar array, so while updating answer array, you apply the f(x)
function to the auxiliar array.